// src/api/model/contactSnapshotDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export interface ContactSnapshotDto {
  name?: string;
  ico?: string;
  dic?: string;
  country?: string;
  street?: string;
  city?: string;
  psc?: string;
  email?: string;
  phone?: string;
}